Make `run` commands exclusive (#6220)
As part of this, loosened up handling of `exclusive` commands such that
downstream commands not marked as `exclusive` are automatically treated
as such. This is to avoid breakage of anyone who is currently defining
an `override def run() = Task.Command{ super.run()() }`, which would
otherwise start failing since `super.run` has suddenly become
`exclusive`. Only downstream commands are treated as such - non-command
tasks are still an error - and I expect this is an uncommon enough
scenario that it won't affect much

Also cleaned up the messiness in `BspRunModule` and tweaked `bspRun` to
avoid the `Task.Anon` that would interfere with exclusivity

Fix https://github.com/com-lihaoyi/mill/issues/5208