Allow users to provide their own compiler bridge (#6250)
This adds a `ScalaModule#scalaCompilerBridge` task, whose value is an
`Option[PathRef]`. It allows users to pass their own compiler bridge,
instead of having Mill fetch and / or compile one on its own internally.

This is useful when trying to build Scala 3 with Mill: Scala 3 builds
its own bridge, that we want to use at some point in the Scala 3 build.