Fix API leakage in Sonatype publisher (#6256)
Fix https://github.com/com-lihaoyi/mill/issues/6251

This fix deliberately breaks binary compatibility of the class
`SonatypeCenrtralPublisher` since the replaced API uses types from an
internal dependency typically only availiable at runtime. It replaces
previously leaked API by new classes `publish.SonatypeCredentials` and
`publish.PublishingType`.

Why I think it is safe to break API:
* Potentially affected users of the breaking API must have worked around
this issues before, since the exposed types in the API where not
available on the compile classpath. But we never got any issues reported
(except the one we try to fix here)
* The API we break is most likely only used in build scripts, which will
be re-compliled with any Mill version change anyways and are therefore
not affected

The alternative fix to the initial issue is to expose the leaked
sonatype central client API as part of the Mill API. This API isn't
stable yet, so there is a risk of breakages later, that are not in our
control.

Pull request: https://github.com/com-lihaoyi/mill/pull/6256