Convert `Result.Failure` into a more structured representation (#6278)
Rather than just being a dumb wrapper of `error: String`, it now
basically is a `List[(error: String, path: os.Path = null, index: Int =
-1)]`. Failures can then be combined with `Failure.combine`, and this
lets us defer rendering until "later" so that the code reporting a
failure doesn't need to be aware of the intricacies of formatting and
rendering errors.

For backwards-compatibility purposes this is implemented as an intrusive
linked list, with the `path` and `index` fields optional and an optional
`next: Option[Failure]` field pointing to the next one. Someone who uses
`Failure` naively as before would simply get the first `error` while
being unaware of the other fields, which is probably fine.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>