Fix codesig detection for overriden tasks (#6290)
Previously we were incorrectly looking up the signature of the
overridden task via it's concrete subclass, which meant that changes in
the overridden superclass task were ignored even if it was used via
`super`. This PR instead makes use of the last `Segment.label` of
overridden tasks which contains a unique suffix of the defining class
name, and uses that instead.

Covered by an integration test

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>