Remove `--add-default` parameter from `testForked` due to breaking semantic compatibility (#6298)
This is causing problems rebootstrapping due to the `override def
testForked` in `integration/package.mill` no longer working due to the
change in signature.

It seems that for non-final tasks, adding parameters cannot be done
safely in a binary/semantic-compatible way. That is the same limitation
we hit in
https://docs.scala-lang.org/sips/unroll-default-arguments.html, and I
guess is just something we'll have to live with