Extend Native (GraalVM) support (#6315)
## Motivation

- Related to https://github.com/com-lihaoyi/mill/issues/6310
- Upon investigation I've discovered we can extend native support to
multiple libraries using
https://github.com/oracle/graalvm-reachability-metadata which is what
both gradle and maven graal vm plugins are doing
- The Kotlin example is not using the maven layout per @lefou 's
suggestion

## Implementation

- For every direct mvn dependency, we get the relevant directory from
the graalvm-reachability-metadata.
- The metadata need to both be in run classpath and passed as a
`--configurations-path`
- The `--configurations-path` is used to simplify native image options -
no more scraping the properties file
- The maven deps use the resolution to get all runtime dependencies and
resolve their reachability metadata
- Both finding the reachability metadata and copying them over reuse the
native-build-tools functionality which uses a walk file tree logic
compatible to the metadata structure.

## Outcomes + examples

- To demonstrate this working, there's an example with Kotlin + Spring
Boot + native configuration which before was failing with
```
141] kotlin.jvm.KotlinReflectionNotSupportedError: Kotlin reflection implementation is not found at runtime. Make sure you have kotlin-reflect.jar in the classpath
```
The error was because we didn't build the native image with the
reachability metadata.

- I've also added some unit tests to show some "unexpected" behaviour
and the need to reuse the copying mechanism from native-build-tools
instead of copying on our own, as it takes care of how the structure is
in graalvm-reachability-metadata. For more information on this
functionality see
[here](https://github.com/graalvm/native-build-tools/blob/master/common/graalvm-reachability-metadata/src/main/java/org/graalvm/reachability/DirectoryConfiguration.java#L95)

Pull request: https://github.com/com-lihaoyi/mill/pull/6315