Ensure zinc worker process is re-used for consecutive compiles (#6349)
Previously the zinc worker was closed pretty aggressively, because
`RefCountedCache` closes it immediately upon release. This PR puts a
`CachedFactoryWithInitData` in front of the `RefCountedCache` to keep
things alive a bit longer, mirroring the two-layer caching we have for
classloaders in various places.

The layered caching setup probably isn't ideal, but it gets tests
passing for now and we can tweak it later

Covered by an integration test