Ensure YAML processing logic preserves original string text (#6382)
To do this, we take all `ScalarNode`s parsed by `org.snakeyaml.engine`
and pass them to uPickle via `visitString`, ignoring their tag and never
using uPickle's other `visitTrue`/`visitFalse`/`visitFloat*`/etc.
methods.

This ensures that the raw input text is preserved for downstream parsers
that require it, and all the built-in `upickle.Reader`s which parse
primitive types are able to gracefully handle `visitString` anyway and
will parse and coerce it as necessary.