Warn if non-platform dependencies are used (#6408)
In the presence of a configured non-empty `platformSuffix`, using
non-platform dependencies is often an error, e.g. for Scala.js or Scala
Native. This can lead to compile and runtime issues, hard to spot and
diagnosem the culprit is often just a single missing colon (`:`).

Mill now prints a warning at resolution time, when a non-platform
dependency is used, but a platform suffix is present. Users can
explicitly disable the warning via the `resolvedDepsWarnNonPlatform`
task.

`MillBuildRootModule` also sets a `platfromSuffix`, to properly resolve
Mill plugins, but we disable the warning here, since using normal jars
is the typical use case for `MillBuildRootModule.mvnDeps`.

One downside to warn at resolution time is, that resolution is cached,
and the warning will not be re-shown if the user does not change the
dependencies. Should we move this to other places, e.g. before `run` or
`test`?

Pull request: https://github.com/com-lihaoyi/mill/pull/6408