Fix worker setup in JvmWorkerModule
The `worker` and `internalWorker` worker tasks were referring
to the same worker instance. As both inherit `AutoClosable`
but only one effective instance is created, there is a risk of
calling `close()` twice and even worse, there is a risk of calling
`close()` on one instance, while the other one is still in use.