Fix worker setup in JvmWorkerModule (#6430)
The `worker` and `internalWorker` worker tasks were referring to the
same worker instance. As both inherit `AutoClosable` but only one
effective instance is created, there is a risk of calling `close()`
twice and even worse, there is a risk of calling `close()` on one
instance, while the other one is still in use.

We fix this by introducing a delegating facade, that does nothing when
closed (by not implementing `AutoClosable`) and just forwards all
invocations to the internal worker.

Pull request: https://github.com/com-lihaoyi/mill/pull/6430